improved portal
As stated in the delete, this is the portal, but split into different files to increase readability.
There are also a few new features among this commit, which are as followed:
added a Sign-in page (only design)
added a Sign-up page(only design)
added a sort function so submit now does something
Changed how the program determines the position of new items. (from row position - 1 to previous row position + 1),  This allows the numbers to stay as low as they can and also prevents duplicates.
added basic navigation features around the files.